Moving TBitHelper and TEscCharsetConverter into Util\Helpers folder
We could do the same with TUtfConverter, but that might require people to change their code on upgrade.

TBitHelper and TEscCharsetConverter are new.
