+++ framework/classes.php
@@ -282,6 +282,8 @@ return [
 'TShellCronLogBehavior' => 'Prado\Util\Cron\TShellCronLogBehavior',
 'TShellDbCronAction' => 'Prado\Util\Cron\TShellDbCronAction',
 'TTimeScheduler' => 'Prado\Util\Cron\TTimeScheduler',
+'TBitHelper' => 'Prado\Util\Helpers\TBitHelper',
+'TEscCharsetConverter' => 'Prado\Util\Helpers\TEscCharsetConverter',
 'IBaseBehavior' => 'Prado\Util\IBaseBehavior',
 'IBehavior' => 'Prado\Util\IBehavior',
 'IClassBehavior' => 'Prado\Util\IClassBehavior',
@@ -294,7 +296,6 @@ return [
 'TBaseBehavior' => 'Prado\Util\TBaseBehavior',
 'TBehavior' => 'Prado\Util\TBehavior',
 'TBehaviorsModule' => 'Prado\Util\TBehaviorsModule',
-'TBitHelper' => 'Prado\Util\TBitHelper',
 'TBrowserLogRoute' => 'Prado\Util\TBrowserLogRoute',
 'TCallChain' => 'Prado\Util\TCallChain',
 'TClassBehavior' => 'Prado\Util\TClassBehavior',
@@ -304,7 +305,6 @@ return [
 'TDbParameterModule' => 'Prado\Util\TDbParameterModule',
 'TDbPluginModule' => 'Prado\Util\TDbPluginModule',
 'TEmailLogRoute' => 'Prado\Util\TEmailLogRoute',
-'TEscCharsetConverter' => 'Prado\Util\TEscCharsetConverter',
 'TFileLogRoute' => 'Prado\Util\TFileLogRoute',
 'TFirebugLogRoute' => 'Prado\Util\TFirebugLogRoute',
 'TFirePhpLogRoute' => 'Prado\Util\TFirePhpLogRoute',
